Add files via upload
This script takes an HTRF readout file (384-well), assuming concentration of controls and blanks in column 23 and 24, and interpolates concentrations of samples quad-stamped (Z-format) into 384-well. 4-point stamp with quad-mapping performed to account for dynamic range of HTRF assays. Dilution factors are incorporated into script, but can be updated to account for changes in protocol. Additionally, the positive control standard concentrations are hardcoded, but can be updated to accommodate changes in protocol.